fix move ClickHandler to element.click package
