+++ kotlin/src/main/kotlin/org/incendo/interfaces/kotlin/paper/PaperExt.kt
@@ -8,7 +8,7 @@ import org.incendo.interfaces.core.arguments.InterfaceArgument
 import org.incendo.interfaces.core.pane.Pane
 import org.incendo.interfaces.core.view.InterfaceView
 import org.incendo.interfaces.paper.PlayerViewer
-import org.incendo.interfaces.paper.element.ClickHandler
+import org.incendo.interfaces.paper.element.click.ClickHandler
 import org.incendo.interfaces.paper.element.ItemStackElement
 import org.incendo.interfaces.paper.type.ChestInterface
 import org.incendo.interfaces.paper.type.TitledInterface
